fixed some signedness problems

+++ debug.c
@@ -218,7 +218,7 @@ dsrv_log(log_t level, char *format, ...) {
   va_list ap;
   FILE *log_fd;
 
-  if (maxlog < level)
+  if (maxlog < (int) level)
     return;
 
   log_fd = level <= DTLS_LOG_CRIT ? stderr : stdout;
@@ -300,7 +300,7 @@ dtls_dsrv_hexdump_log(log_t level, const char *name, const unsigned char *buf, s
   FILE *log_fd;
   int n = 0;
 
-  if (maxlog < level)
+  if (maxlog < (int) level)
     return;
 
   log_fd = level <= DTLS_LOG_CRIT ? stderr : stdout;
